requirements.txt
esptool==2.8
wget==3.2
python-socketio~=4.6.0
sanic~=20.6.3
sanic_compress
python-socketio
aiofile~=3.0.0
python-dateutil~=2.8.1
aiohttp~=3.6.2
requests~=2.22.0
pyserial~=3.4
